Mostly working TAC generation for small main functions

+++ test_files/test9.c
@@ -1,7 +1,6 @@
-// Parser , Should cause Type error
-int main() {
-    int x = 5;
+float main() {
+    float x = 5.0;
     float y = 3.14;
-    int z = x + y; // int + float
+    float z = x + y; 
     return z;
 }
